+++ event-data-model/test/Main.hs
@@ -1,8 +1,0 @@
-module Main
-  ( main
-  ) where
-
-import           EventData.Tests
-main :: IO ()
-main = do
-  edmTests